adding places endpoint for strech task
